+++ Realm/NSError+RLMSync.h
@@ -1,6 +1,6 @@
 ////////////////////////////////////////////////////////////////////////////
 //
-// Copyright 2014 Realm Inc.
+// Copyright 2017 Realm Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,24 +18,24 @@
 
 #import <Foundation/Foundation.h>
 
-#import <Realm/RLMArray.h>
-#import <Realm/RLMMigration.h>
-#import <Realm/RLMObject.h>
-#import <Realm/RLMObjectSchema.h>
-#import <Realm/RLMPlatform.h>
-#import <Realm/RLMProperty.h>
-#import <Realm/RLMRealm.h>
-#import <Realm/RLMRealmConfiguration.h>
-#import <Realm/RLMRealmConfiguration+Sync.h>
-#import <Realm/RLMResults.h>
-#import <Realm/RLMSchema.h>
-#import <Realm/RLMSyncConfiguration.h>
-#import <Realm/RLMSyncCredentials.h>
-#import <Realm/RLMSyncManager.h>
-#import <Realm/RLMSyncPermissionChange.h>
-#import <Realm/RLMSyncPermissionOffer.h>
-#import <Realm/RLMSyncPermissionOfferResponse.h>
-#import <Realm/RLMSyncSession.h>
-#import <Realm/RLMSyncUser.h>
-#import <Realm/RLMSyncUtil.h>
-#import <Realm/NSError+RLMSync.h>
+NS_ASSUME_NONNULL_BEGIN
+
+@interface NSError (RLMSync)
+
+/**
+ Given a Realm Object Server client reset error, return the block that can
+ be called to manually initiate the client reset process, or nil if the
+ error isn't a client reset error.
+ */
+- (nullable void(^)(void))rlmSync_clientResetBlock NS_REFINED_FOR_SWIFT;
+
+/**
+ Given a Realm Object Server client reset error, return the path where the
+ backup copy of the Realm will be placed once the client reset process is
+ complete.
+ */
+- (nullable NSString *)rlmSync_clientResetBackedUpRealmPath NS_SWIFT_UNAVAILABLE("");
+
+@end
+
+NS_ASSUME_NONNULL_END