+++ Realm/RLMSyncConfiguration.mm
@@ -35,7 +35,9 @@ namespace {
 using ProtocolError = realm::sync::ProtocolError;
 
 RLMSyncSystemErrorKind errorKindForSyncError(SyncError error) {
-    if (error.error_code == ProtocolError::bad_authentication) {
+    if (error.is_client_reset_requested()) {
+        return RLMSyncSystemErrorKindClientReset;
+    } else if (error.error_code == ProtocolError::bad_authentication) {
         return RLMSyncSystemErrorKindUser;
     } else if (error.is_session_level_protocol_error()) {
         return RLMSyncSystemErrorKindSession;
@@ -140,11 +142,16 @@ static BOOL isValidRealmURL(NSURL *url) {
             errorHandler = [=](std::shared_ptr<SyncSession> errored_session,
                                SyncError error) {
                 RLMSyncSession *session = [[RLMSyncSession alloc] initWithSyncSession:errored_session];
+                NSMutableDictionary *userInfo = [NSMutableDictionary dictionaryWithCapacity:error.user_info.size()];
+                for (auto& pair : error.user_info) {
+                    userInfo[@(pair.first.c_str())] = @(pair.second.c_str());
+                }
                 // FIXME: how should the binding respond if the `is_fatal` bool is true?
                 [[RLMSyncManager sharedManager] _fireErrorWithCode:error.error_code.value()
                                                            message:@(error.message.c_str())
                                                            isFatal:error.is_fatal
                                                            session:session
+                                                          userInfo:userInfo
                                                         errorClass:errorKindForSyncError(error)];
             };
         }