+++ Realm/RLMSyncSession_Private.hpp
@@ -27,7 +27,10 @@ class SyncSession;
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface RLMSyncSession () RLM_SYNC_UNINITIALIZABLE
+@interface RLMSyncSession () {
+@public     // So it's visible to tests
+    std::weak_ptr<realm::SyncSession> _session;
+} RLM_SYNC_UNINITIALIZABLE
 
 - (instancetype)initWithSyncSession:(std::shared_ptr<realm::SyncSession>)session;